implement aligned fill, but layout convergence is currently a (theoretical) problem:
Alexandria is only finished after four layout iterations. Since the bibliography and list of prompts (and their respective headings) are only rendered conditionally based on whether the bibliography is empty, that means that the full list of outline entries is only ready in the fifth iteration. That means the fill alignment function can only _measure_ correctly in the fifth iteration, and would thus require a sixth iteration for _rendering_

as long as the bibliographies don't push the document from 99 to 100, the nonconvergence is actually not a problem, but this still needs to be fixed.